agents/requirements.txt
strands-agents==0.1.7
strands-agents-tools==0.1.5